requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
avro-python3==1.10.2
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==5.1.1
duckdb==1.0.0
frozenlist==1.4.1
fsspec==2024.6.0
gcsfs==2024.6.0
google-api-core==2.19.0
google-api-python-client==2.128.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
numpy==1.26.4
oauthlib==3.2.2
pandas==2.2.2
pandasql==0.7.3
proto-plus==1.23.0
protobuf==4.25.3
py4j==0.10.9.7
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
pyspark==3.5.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
sendgrid==6.11.0
simple_webbrowser==0.2.2
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.30
starkbank-ecdsa==2.2.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
yarl==1.9.4
